+++ R/read_aircrafts.R
@@ -30,6 +30,7 @@ read_aircrafts <- function(date = 202001,
 ### check inputs
   if( ! is.logical(showProgress) ){ stop(paste0("Argument 'showProgress' must be either 'TRUE' or 'FALSE.")) }
   if( ! is.logical(cache) ){ stop(paste0("Argument 'cache' must be either 'TRUE' or 'FALSE.")) }
+  check_input_date_format(date)
 
   ### check date input
   # get all dates available